back/requirements.txt
Django==4.0.4
psycopg2-binary==2.9.3
django-cors-headers==3.12.0
djangorestframework==3.13.1
django-allauth==0.50.0
dj-rest-auth==2.2.4
djangorestframework-simplejwt==5.1.0